requirements.txt
absl-py==0.15.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
asttokens==2.0.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
black==21.9b0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cytoolz==0.11.0
dataclassy==0.11.1
einops==0.3.2
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.1
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
fsspec==2021.11.0
ftfy==6.0.3
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
hexbytes==0.2.2
hypothesis==6.24.0
idna==3.3
imageio==2.10.3
imageio-ffmpeg==0.4.5
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
kornia==0.6.1
lazy-object-proxy==1.6.0
lru-dict==1.1.7
Markdown==3.3.4
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
numpy==1.21.4
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.0
parsimonious==0.8.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.21
protobuf==3.19.0
psutil==5.8.0
py==1.10.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
pyDeprecate==0.3.1
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==3.0.3
pyrsistent==0.18.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.6.1
pytorch-lightning==1.5.0
pywin32==302
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rlp==2.0.1
rsa==4.7.2
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
taming-transformers==0.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.1
torch==1.10.0
torchmetrics==0.6.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
vvm==0.1.0
vyper==0.3.0
wcwidth==0.2.5
web3==5.24.0
websockets==9.1
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.2
yarl==1.7.0
